Enable control of GNSS antenna
Turn on active antenna voltage control, short detection, open
detection and enable power down in the event of a short circuit.
Implement the features by writing boolean true into each of 4
registers at one go, which requires implementation of a new message
type.  Enabling only one of the features, or disabling them, is not
supported, but would be a straightforward extension of the current
code.

The hardware's acknowledgement messages which confirm some of the
configuration change are visible in the output:

root@medusa:~# cat /dev/gnss0
[ . . . ]
$GNTXT,01,01,02,ANTSUPERV=AC SD OD PDoS *14
$GNTXT,01,01,02,ANTSTATUS=INIT*3B
$GNTXT,01,01,02,ANTSTATUS=OK*25

On each boot, the first few bytes available from the gnss0 device are
always garbled.   Adding a delay at the end of the open() command does
not make the garbled bytes go away.