GNSS: fix panic on concurrent userspace access
Every time the GNSS device is opened, the reference counts on both the
device and the tty connected to the serial device must increase.  When
the device is closed, the reference counts must be correspondingly
decremented.  If the reference count is wrong, then a userspace
process which accesses the GNSS may double-free or leak the file
descriptor.  The warning

ttyS ttyS1: tty_port_close_start: tty->count = 1 port count = 2

indicates that the serial port and GNSS reference counts are out of
sync.

The TTY's reference counters are controlled via the pm_runtime_*
functions, but the GNSS reference counters accessible to userspace via
sysfs rely on the get_device() and put_device() commands.  Protect the
counters with the GNSS device semaphore in the case of concurrent
userspace accesses.  Pull the configuration commands out into their
own function to improve readability.

Jira-issue: AV-267804