GNSS: configure the device at driver probe
GNSS users reasonably expect the device driver to configure the
antenna and PPS before the device node is opened.  The only
opportunity to do so is from the driver probe method.  Configuring the
settings requires opening the serial device, which is odd from inside
the probe method.  Nonetheless, the configuration should work, as
configuration code writes to the underlying serial device, not to the
device node which the probe() method creates.

With the patch, at boot:

medusa-cm login: root
root@medusa-cm:~# dmesg | grep -i gnss
[    2.170680] gnss: GNSS driver registered with major 249
[    2.170802] gnss-ubx serial0-0: supply vcc not found, using dummy regulator
[    2.202819] gnss gnss0: Set GNSS speed to 921600
[    2.202838] gnss gnss0: Enabling antenna controls.
[    2.214766] gnss gnss0: Enabled GNSS antenna controls.
[    2.214789] gnss gnss0: Selecting UBX output protocol.
[    2.230473] gnss gnss0: Set GNSS protocol to UBX.
[    2.242762] gnss gnss0: Set time pulse reference to GPS.