route clock solution and timing pulse time data to UART1
The UBX_NAV_CLOCK and UBX_TIM_TP messages are not output by default.
Configure the device driver to send them to UART1 along with the other
GNSS messages.  With this code change, the RECEIVED_TIM_TP and
RECEIVED_NAV_CLOCK counters in the GnssStatus messages increment
upward and MISSING_NAV_CLOCK is zero.